refactor(tasks): move G1 shared base owner [skip ci]
