--- src/unilab/envs/locomotion/g1/__init__.py
@@ -1,1 +1,0 @@
-"""Legacy G1 shared base pending motion-tracking migration."""
