+++ src/unilab/tasks/motion_tracking/x2/flip_tracking.py
@@ -8,7 +8,7 @@ from typing import Literal
 from unilab.assets import ASSETS_ROOT_PATH
 from unilab.base import registry
 from unilab.base.scene import SceneCfg
-from unilab.envs.locomotion.g1.base import Sensor
+from unilab.tasks.locomotion.g1.base import Sensor
 
 from ..common.config import (
     PoseRandomization,
